+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- **Breaking**: The `WakuMessage` APIs have been changed to move `contentTopic` out of the optional parameters.
+
+### Removed
+- Examples (web-chat): Remove broken `/fleet` command.
+- **Breaking**: Removed `DefaultContentTopic` as developers must choose a content topic for their app;
+  recommendations for content topic can be found at https://rfc.vac.dev/spec/23/.
+
 ## [0.9.0] - 2021-07-26
 
 ### Changed